requirements.txt
openai
sounddevice
numpy
soundfile
websocket-client
